requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
python-dotenv>=1.0
dj-database-url>=2.1
Pillow>=10.0
qrcode[pil]>=7.4
numpy>=1.26
scikit-learn>=1.5
psycopg[binary]>=3.2
